Fixed AspectJ weaving for plugins
