+++ gradle-plugin/src/main/java/io/fairyproject/gradle/aspect/AspectJCompiler.java
@@ -1,4 +1,4 @@
-package io.fairyproject.gradle.aspectj;
+package io.fairyproject.gradle.aspect;
 
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
