+++ gradle-plugin/src/main/java/io/fairyproject/gradle/aspect/DefaultWeavingSourceSet.java
@@ -1,4 +1,4 @@
-package io.fairyproject.gradle.aspectj;
+package io.fairyproject.gradle.aspect;
 
 import lombok.Data;
 import org.gradle.api.file.FileCollection;
